🎨 Little refacto
