+++ packages/applications/document-builder/src/candidature/attestation/helpers/getDésignationCatégorie.ts
@@ -1,0 +1,20 @@
+import { AppelOffre } from '@potentiel-domain/appel-offre';
+
+export const getDésignationCatégorie = ({
+  puissance,
+  note,
+  periodeDetails,
+}: {
+  puissance: number;
+  note: number;
+  periodeDetails: AppelOffre.Periode;
+}) => {
+  if (periodeDetails.noteThresholdBy !== 'category') {
+    return;
+  }
+
+  return puissance <= periodeDetails.noteThreshold.volumeReserve.puissanceMax &&
+    note >= periodeDetails.noteThreshold.volumeReserve.noteThreshold
+    ? 'volume-réservé'
+    : 'hors-volume-réservé';
+};
